Use sbyte for card raw value instead of short

+++ src/CommandLineHelper.cs
@@ -204,7 +204,7 @@ namespace FreeCellSolver.Entry
                     await WriteResourceAsync("FreeCellSolver.visualizer.dist.index.min.js", Path.Combine(path, "index.min.js"));
                     await WriteResourceAsync("FreeCellSolver.visualizer.dist.visualizer.min.html", Path.Combine(path, "visualizer.html"));
 
-                    for (short i = 0; i < 52; i++)
+                    for (sbyte i = 0; i < 52; i++)
                     {
                         Card.Get(i).ToImage().Save(Path.Combine(path, "assets", $"{i}.png"));
                     }
